d4n/policy: Change `size_t` argument type and add error checking to `set_local_weight_sum` method

+++ src/rgw/driver/d4n/d4n_policy.h
@@ -72,7 +72,7 @@ class LFUDAPolicy : public CachePolicy {
 
     int set_age(int age, optional_yield y);
     int get_age(optional_yield y);
-    int set_local_weight_sum(size_t weight, optional_yield y);
+    int set_local_weight_sum(int weight, optional_yield y);
     int get_local_weight_sum(optional_yield y);
     CacheBlock* get_victim_block(const DoutPrefixProvider* dpp, optional_yield y);
